Reuse toString for all Tuple derived classes
